+++ src/service/ExpressionParser.js
@@ -84,7 +84,8 @@ export default class ExpressionParser {
     const parser = Object.getOwnPropertyNames( this.parsers )
       .find( ( available ) => exp.startsWith( available + "(" ) );
     if ( !parser ) {
-      throw new ExpressionParserException( `Cannot parse expression ${ exp }. Expected syntax {{ method(..) }}` );
+      console.warn( `Cannot parse expression ${ exp }. Expected syntax {{ method(..) }}` );
+      return `\`${ rawExp }\``;
     }
     try {
       return this.parsers[ parser ]( extractParams( parser, exp ) );